refactoring agent service
